Replacing AsyncHttp Client with more advanced OKHttp.

+++ app/src/main/java/com/fetherz/flicks/models/movie/Movie.java
@@ -49,7 +49,7 @@ public class Movie implements Serializable {
     double popularity;
 
     @SerializedName("vote_count")
-    double voteCount;
+    int voteCount;
 
     @SerializedName("video")
     boolean isVideo;
@@ -107,7 +107,7 @@ public class Movie implements Serializable {
         }
     }
 
-    public double getVoteCount() {
+    public int getVoteCount() {
         return voteCount;
     }
